[change]motion.cの中身をほぼ削除
現在の機体（エンコーダ改修済み）で直進するプログラムを入れてあります

+++ src/peripheral/motor.h
@@ -1,27 +1,34 @@
 #ifndef MOTOR_H
 #define MOTOR_H
 
+#include <stdint.h>
+
 /**
  * モータPWMを発生させるためのTimer0初期設定を行います．
  */
 void init_timer0(void);
+
 /**
  * モータPID制御を行うためのTimer2初期設定を行います．
- * これにより，割り込み関数{@code ISR(TIMER2_OVF_vect)}は、約13msec周期(76.3Hz)で呼びだされます．
- * また，PID制御については，
+ * これにより，割り込み関数{@code ISR(TIMER2_OVF_vect)}は、約13msec周期(約76.3Hz)で呼びだされます．
+ * ただし、{@code set_t2intr_callback}で登録した関数はmotor.cで定義された回数に1回だけ呼ばれます。
  */
 void init_timer2(void);
+
 /**
  * ロータリエンコーダ割り込みを制御するための初期化を行います．
  */
 void init_intr(void);
+
 /**
  * Timer2割り込み時にコールバックする関数(引数は右左のロータリエンコーダのカウント値)を指定します．
  */
-void set_t2intr_callback(void (*f)(int8_t, int8_t));
+void set_t2intr_callback(void (*f)(int, int));
+
 /**
  * モータに出力するPWMの値を設定します．
  * @param r_pwm 右モータに入力するPWMの値()
+ * @param l_pwm 左モータに入力するPWMの値()
  */
 void set_motor_pwm(int16_t r_pwm, int16_t l_pwm);
